+++ Changes.md
@@ -16,6 +16,7 @@ Fixes
   - PxrSurface, PxrLayerSurface : Moved `utilityPattern` parameter to "Globals" section.
   - Fixed formatting of parameter tooltips, and Node Reference descriptions.
 - RenderManShader : Fixed loading of C++ pattern shaders such as `aaOceanPrmanShader`.
+- ArnoldShader : Fixed `correspondingInput()` to handle children of the main output plug.
 
 API
 ---